automation/requirements.txt
pandas
scikit-learn
reportlab
google-generativeai
python-dotenv
schedule
